backend/services/scraper/requirements.txt
python-dotenv
selenium
pandas
pyarrow
pymongo
schedule